Add files via upload
Medical Imaging (CT-Xray) Colorization New Dataset 🩺💻🖼️ This dataset provides a collection of medical imaging data, including both CT (Computed Tomography) and X-ray images, with an added focus on colorization techniques. The goal of this dataset is to facilitate the enhancement of diagnostic processes by applying various colorization techniques to grayscale medical images, allowing researchers and machine learning models to explore the effects of color in radiology.